Add components folder and update file paths
Fix oauth that was lost to git merge
