+++ src/Components/Sort/Sort.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Card from './Card.js';
+import Card from '../Card/Card.js';
 
 export default class Sort extends Component {
   constructor() {
